backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
langchain
python-dotenv
